added todaysStats component calendar styles
